Implemented id in autoincrement
This commit implements the use of an auto-incrementing primary key for the report_id field in the reports table migration. By defining report_id as a big integer with auto-increment, each new report record will automatically receive a unique identifier. Additionally, the migration establishes a foreign key relationship between the super_admin_id in the reports table and the super_admin_id in the super_admins table, ensuring referential integrity and enabling cascading deletes. These changes improve the structure and reliability of the database schema for managing report records.